requirements.txt
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-oauthlib>=0.8.0
python-dotenv>=1.0.0
google-generativeai>=0.1.0
tenacity>=8.2.0
